syncval: Segregate AccessContext initial pass
AccessContext class and method into their own files.

Need to clean up RenderPass/AccessContext interface and segregated
