+++ BUILD.gn
@@ -237,6 +237,8 @@ vvl_sources = [
   "layers/stateless/sl_utils.cpp",
   "layers/stateless/sl_wsi.cpp",
   "layers/stateless/stateless_validation.h",
+  "layers/sync/sync_access_context.cpp",
+  "layers/sync/sync_access_context.h",
   "layers/sync/sync_common.h",
   "layers/sync/sync_model.cpp",
   "layers/sync/sync_model.h",
